requirement.txt
beautifulsoup4==4.12.2
blis==0.7.9
bs4==0.0.1
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
cssselect==1.2.0
cymem==2.0.7
decorator==5.1.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.5.0/en_core_web_trf-3.5.0-py3-none-any.whl
et-xmlfile==1.1.0
extruct==0.14.0
filelock==3.12.2
fsspec==2023.6.0
future==0.18.3
geocoder==1.38.1
html-text==0.5.2
html5lib==1.1
huggingface-hub==0.15.1
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
jstyleson==0.0.2
langcodes==3.3.0
lxml==4.9.2
MarkupSafe==2.1.3
mf2py==1.1.2
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.1
numpy==1.25.0
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
pathy==0.10.1
phonenumbers==8.13.14
preshed==3.0.8
probableparsing==0.0.1
pyap==0.3.1
pydantic==1.10.9
pyparsing==3.1.0
pyRdfa3==3.5.3
python-crfsuite==0.9.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
ratelim==0.1.6
rdflib==6.3.2
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
six==1.16.0
smart-open==6.3.0
soupsieve==2.4.1
spacy==3.5.3
spacy-alignments==0.9.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-transformers==1.2.5
srsly==2.4.6
sympy==1.12
thefuzz==0.19.0
thinc==8.1.10
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.30.2
typer==0.7.0
typing-extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
usaddress==0.5.10
usaddress-scourgify==0.5.1
w3lib==2.1.1
wasabi==1.1.2
webencodings==0.5.1
yaml-config==0.1.5
